// src/common/enums.ts
export enum UserRole {
  STUDENT = 'STUDENT',
  TEACHER = 'TEACHER',
  LIBRARIAN = 'LIBRARIAN',
  ADMIN = 'ADMIN',
}

export enum ItemStatus {
  AVAILABLE = 'AVAILABLE',
  BORROWED = 'BORROWED',
  RESERVED = 'RESERVED',
  LOST = 'LOST',
  DAMAGED = 'DAMAGED',
  MAINTENANCE = 'MAINTENANCE',
}

export enum ItemType {
  BOOK = 'BOOK',
  EBOOK = 'EBOOK',
  AUDIOBOOK = 'AUDIOBOOK',
  DVD = 'DVD',
  BLURAY = 'BLURAY',
  CD = 'CD',
  MAGAZINE = 'MAGAZINE',
  NEWSPAPER = 'NEWSPAPER',
  JOURNAL = 'JOURNAL',
  THESIS = 'THESIS',
  REFERENCE = 'REFERENCE',
  MAP = 'MAP',
  GAME = 'GAME',
  SOFTWARE = 'SOFTWARE',
  EQUIPMENT = 'EQUIPMENT',
}

export enum LoanStatus {
  BORROWED = 'BORROWED',
  PENDING_RETURN = 'PENDING_RETURN',
  RETURNED = 'RETURNED',
}

export enum Genre {
  FICTION = 'FICTION',
  NON_FICTION = 'NON_FICTION',
  SCIENCE = 'SCIENCE',
  HISTORY = 'HISTORY',
  TECHNOLOGY = 'TECHNOLOGY',
  ART = 'ART',
  OTHER = 'OTHER',
}

export enum ActivityType {
  LOAN_CREATED = 'LOAN_CREATED',
  LOAN_RETURNED = 'LOAN_RETURNED',
  LOAN_RENEWED = 'LOAN_RENEWED',
  RESERVATION_PLACED = 'RESERVATION_PLACED',
  RESERVATION_CANCELLED = 'RESERVATION_CANCELLED',
  FINE_APPLIED = 'FINE_APPLIED',
  FINE_PAID = 'FINE_PAID',
  FINE_WAIVED = 'FINE_WAIVED',
  USER_SUSPENDED = 'USER_SUSPENDED',
  USER_ACTIVATED = 'USER_ACTIVATED',
  ITEM_ADDED = 'ITEM_ADDED',
  ITEM_UPDATED = 'ITEM_UPDATED',
  ITEM_ARCHIVED = 'ITEM_ARCHIVED',
  LIBRARY_SETTINGS_CREATED = 'LIBRARY_SETTINGS_CREATED',
  LIBRARY_SETTINGS_UPDATED = 'LIBRARY_SETTINGS_UPDATED',
  LIBRARY_SETTINGS_RESET = 'LIBRARY_SETTINGS_RESET',
  USER_REGISTERED = 'USER_REGISTERED',
  USER_LOGIN = 'USER_LOGIN',
  USER_LOGOUT = 'USER_LOGOUT',
  EMAIL_VERIFIED = 'EMAIL_VERIFIED',
  PASSWORD_RESET_REQUESTED = 'PASSWORD_RESET_REQUESTED',
  PASSWORD_RESET_COMPLETED = 'PASSWORD_RESET_COMPLETED',
  ADMIN_USER_CREATED = 'ADMIN_USER_CREATED',
  ITEM_BORROWED = 'ITEM_BORROWED',
  ITEM_RETURNED = 'ITEM_RETURNED',
  PROFILE_UPDATED = 'PROFILE_UPDATED',
  ROLE_CHANGED = 'ROLE_CHANGED',
}